Restrict most IAccessible event processing to objects of interest to the user.
NVDA already dropped events for background windows, etc.
However, there can be a lot of events even in the foreground window that aren't important to the user.
For example, a nameChange on an object that isn't focused and isn't in the focus ancestry isn't reported in any way.
Previously, these events were processed, NVDAObjects were instantiated for them and the events were queued, but they simply did nothing when they were finally handled by the NVDAObject.
This could cause noticeable performance problems where a lot of events are fired, particularly on the web.
To address this, only process most IAccessible events for the focus, focus ancestry, navigator object and desktop object.

Implementation:

1. Add NVDAObject.matchAPIEvent, which takes parameters specific to the platform API (win event parameters for IAccessible) and returns True if the event is for this object.
2. eventHandler.shouldAcceptEvent now takes an optional apiEventParams argument.
    If supplied, it checks whether the event is for any of the NVDAObjects of interest to the user.
    If not (and the event wasn't accepted by some other rule), the event is dropped.
3. IAccessibleHandler now calls shouldAcceptEvent with appropriate apiEventParams.
    The call to shouldAcceptEvent has been moved into winEventCallback so we drop events as early as possible.
    Focus events are still checked in pumpAll, since checking those earlier caused problems in the past.